+++ R/misc.R
@@ -12,10 +12,8 @@
 #' require_pkgs(c("a", "base", "b", "c"))
 #' }
 #'
-#' @importFrom purrr
-#' map_lgl
-#' @importFrom assertthat
-#' assert_that
+#' @importFrom purrr map_lgl
+#' @importFrom assertthat assert_that
 #' @keywords internal
 require_pkgs <- function(pkgs) {
   assert_that(is.character(pkgs))
@@ -33,4 +31,24 @@ require_pkgs <- function(pkgs) {
       call. = FALSE
     )
   }
+}
+
+
+#' Check availability of a URL
+#'
+#' @details The code is based on \url{https://stackoverflow.com/a/60627969}.
+#'
+#' @return A logical of length 1.
+#'
+#' @keywords internal
+url_is_up <- function(url, timeout = 2) {
+  con <- url(url)
+  check <- suppressWarnings(
+    try(
+      open.connection(con, open = "rt", timeout = timeout),
+      silent = TRUE
+    )[1]
+  )
+suppressWarnings(try(close.connection(con), silent = TRUE))
+ifelse(is.null(check), TRUE, FALSE)
 }